requirements.txt
thefuzz
openpyxl
semantic-kernel==1.23.1
gradio==5.14.0
requests
pypandoc
python-dotenv
pillow
python-docx
pytest